ConfigJson解析完成，以及细节bug修复，Gui制作

+++ src/main/java/com/aljun/uninfectedzone/core/file/FileUtils.java
@@ -39,9 +39,11 @@ public class FileUtils {
         writer.close();
     }
 
-    public static void deleteJsonFile(String path) {
+    public static void deleteFile(String path) {
         File file = new File(path);
-        file.delete();
+        if (file.exists()) {
+            file.delete();
+        }
     }
